workbooks/05-starter-code/flask-postgresql/requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
psycopg2-binary